Fixed minor error in sample command in main slides.
A sample on application of mv command for renaming files had a minor error. Since the files to be renamed were being taken in a loop and they had spaces in their names, the mv command took the 2 parts of the string '$i' as its 2 inputs.Changed $i to "$i" and it works just fine.